Migrate to SDL3, CMakePresets, and modernize build system
Replace SDL2 with SDL3 (3.4.2) and update all SDL API usage throughout
the codebase. Add CMakePresets.json with default/debug presets so the
vcpkg toolchain is wired automatically via VCPKG_ROOT, eliminating the
custom vcpkg_setup_toolchain.cmake. Update vcpkg baseline to pick up
SDL3 and other current dependency versions.

+++ src/app.hpp
@@ -5,7 +5,7 @@
 #include <memory>
 
 #include <glm/glm.hpp>
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 
 class Window;
 class Context;
@@ -42,7 +42,7 @@ private:
     glm::vec3 resolution_;
     glm::vec3 default_resolution_;
 
-    Uint32 start_time_;
+    Uint64 start_time_;
     bool fullscreen_;
 
     bool high_dpi_;
